backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
joblib
python-dotenv
requests
pydantic
numpy
python-dateutil